validations done

+++ src/app/login/login.component.ts
@@ -18,7 +18,7 @@ export class LoginComponent implements OnInit {
   userType:String;
   userEmail:String = "";
   userPassword:String = "";
-  
+  hide:boolean = true;
   constructor(private service:UserService, private router:Router) { }
 
   onItemChange(value){
@@ -34,16 +34,36 @@ export class LoginComponent implements OnInit {
       this.user.password = this.userPassword;
       this.service.loginUser(this.user).subscribe(
         fetchedData=>{
-          //console.log(fetchedData);
-          localStorage.setItem('customerId', String(fetchedData.userId));
+          console.log(fetchedData);
+          if(fetchedData.userStatus == "Failed")
+          {
+            Swal.fire({
+                 title: "Login Failed",
+                  text:"Please check your credentials",
+                  icon: "error",
+                  confirmButtonText: "Okay"
+              });
+          }
+          else{
+            localStorage.setItem('customerId', String(fetchedData.userId));
+            Swal.fire({
+              title: "Login Success",
+               text:"Please avail services",
+               icon: "success",
+               confirmButtonText: "Okay"
+           });
+           this.router.navigate(['']);
+          }
+          //localStorage.setItem('customerId', String(fetchedData.userId));
           //alert(Number(sessionStorage.getItem('customer')));
-          Swal.fire({
-            title: "Login Successful",
-            text:"Please proceed to avail the services",
-            icon: "success",
-            confirmButtonText: "Okay"
-          });
-          this.router.navigate(['']);
+          
+          // Swal.fire({
+          //   title: "Login Successful",
+          //   text:"Please proceed to avail the services",
+          //   icon: "success",
+          //   confirmButtonText: "Okay"
+          // });
+          // this.router.navigate(['']);
         }
       );
       // this.user.userEmail = this.userEmail;
